+++ .gitignore
@@ -1,0 +1,11 @@
+# dependencies
+/node_modules
+# testing
+/coverage
+# production
+/build
+# misc
+.DS_Store
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
