+++ public/index.html
@@ -14,7 +14,7 @@
 <body>
   <div id="root"></div>
 
-  <script src="script.min.js"></script>
+  <script src="dist/main.bundle.js"></script>
 </body>
 
 </html>
