+++ src/index.js
@@ -1,0 +1,7 @@
+
+
+import { Controller } from "./controller"
+import { Model } from "./model"
+import { View } from "./view"
+
+const app = new Controller(new Model(), new View())
