+++ src/view.js
@@ -1,63 +1,9 @@
 /**
- * @class Model
- *
- * Manages the data of the application.
- */
-class Model {
-  constructor() {
-    this.todos = JSON.parse(localStorage.getItem('todos')) || []
-  }
-
-  bindTodoListChanged(callback) {
-    this.onTodoListChanged = callback
-  }
-
-  _commit(todos) {
-    this.onTodoListChanged(todos)
-    localStorage.setItem('todos', JSON.stringify(todos))
-  }
-
-  addTodo(todoText) {
-    const todo = {
-      id: this.todos.length > 0 ? this.todos[this.todos.length - 1].id + 1 : 1,
-      text: todoText,
-      complete: false,
-    }
-
-    this.todos.push(todo)
-
-    this._commit(this.todos)
-  }
-
-  editTodo(id, updatedText) {
-    this.todos = this.todos.map(todo =>
-      todo.id === id ? { id: todo.id, text: updatedText, complete: todo.complete } : todo
-    )
-
-    this._commit(this.todos)
-  }
-
-  deleteTodo(id) {
-    this.todos = this.todos.filter(todo => todo.id !== id)
-
-    this._commit(this.todos)
-  }
-
-  toggleTodo(id) {
-    this.todos = this.todos.map(todo =>
-      todo.id === id ? { id: todo.id, text: todo.text, complete: !todo.complete } : todo
-    )
-
-    this._commit(this.todos)
-  }
-}
-
-/**
  * @class View
  *
  * Visual representation of the model.
  */
-class View {
+export class View {
   constructor() {
     this.app = this.getElement('#root')
     this.form = this.createElement('form')
@@ -195,50 +141,3 @@ class View {
     })
   }
 }
-
-/**
- * @class Controller
- *
- * Links the user input and the view output.
- *
- * @param model
- * @param view
- */
-class Controller {
-  constructor(model, view) {
-    this.model = model
-    this.view = view
-
-    // Explicit this binding
-    this.model.bindTodoListChanged(this.onTodoListChanged)
-    this.view.bindAddTodo(this.handleAddTodo)
-    this.view.bindEditTodo(this.handleEditTodo)
-    this.view.bindDeleteTodo(this.handleDeleteTodo)
-    this.view.bindToggleTodo(this.handleToggleTodo)
-
-    // Display initial todos
-    this.onTodoListChanged(this.model.todos)
-  }
-
-  onTodoListChanged = todos => {
-    this.view.displayTodos(todos)
-  }
-
-  handleAddTodo = todoText => {
-    this.model.addTodo(todoText)
-  }
-
-  handleEditTodo = (id, todoText) => {
-    this.model.editTodo(id, todoText)
-  }
-
-  handleDeleteTodo = id => {
-    this.model.deleteTodo(id)
-  }
-
-  handleToggleTodo = id => {
-    this.model.toggleTodo(id)
-  }
-}
-
-const app = new Controller(new Model(), new View())
