+++ webpack/config.common.js
@@ -1,0 +1,25 @@
+'use strict';
+
+const path = require('path');
+
+module.exports = {
+
+  entry: {
+    main: path.resolve('./src/index.js')
+  },
+  
+  output: {
+    filename: '[name].bundle.js',
+    path: path.resolve('./public/dist')
+  },
+  
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: 'babel-loader'
+      }
+    ]
+  }
+};
